+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Removed the dependency on the Ruby `benchmark` module.
+
 [Unreleased]: https://github.com/envato/zxcvbn-ruby/compare/v1.2.0...HEAD
 
 ## [1.2.0] - 2021-01-05